Update SlotMachine.py
Its a text based Slot machine. The user will deposit a certain amount of money, where we will allow the to bet on 1 or 2 or 3 lines of slot machine. Then we are going to determine if they won or not. So if they got any lines, we are going to multiply their bet by the value of the line, add that to their balance and allow them to keep playing until they want to cash in or they run out of their money.